Pulling `MockWordPressComApi` into a separate file
It's needed for `PushAuthenticationService` tests.
